snmp alert test has been rewritten
